Refactor partner application handling and remove deprecated benefit icons setup
- Removed the benefit-icons.php helper file and associated setup scripts.
- Consolidated partner application logic into partner-application-handler.php.
- Enhanced nonce verification and duplicate submission prevention using session keys.
- Implemented database locks to prevent race conditions during partner application submissions.
- Streamlined benefit icon retrieval process, including fallback to media library if not found in options.
- Cleaned up functions in api.php for consistency and readability.
- Removed temporary benefit icons setup code from functions.php.
