+++ inc/helpers/api.php
@@ -1,10 +1,12 @@
 <?php
 
-function getUrl($str) {
+function getUrl($str)
+{
   return get_template_directory_uri() . "/$str";
 }
 
-function dump($var, $label = null, $echo = true) {
+function dump($var, $label = null, $echo = true)
+{
   $style = '<style>
   .pretty-dump {
     font-family: monospace;
@@ -48,7 +50,8 @@ function dump($var, $label = null, $echo = true) {
   }
 }
 
-function displayIcon() {
+function displayIcon()
+{
   $arr = get_field('footer_icons', 'option');
 
   if (!is_array($arr)) {
@@ -74,7 +77,8 @@ function displayIcon() {
   }
 }
 
-function getPlaceReviews() {
+function getPlaceReviews()
+{
   $apiKey = get_field('reviews_api_token', 'option');
   $placeId = get_field('reviews_api_place_id', 'option');
   $url = "https://maps.googleapis.com/maps/api/place/details/json?place_id={$placeId}&fields=name,rating,reviews&language=en&key={$apiKey}";
@@ -111,7 +115,8 @@ function getPlaceReviews() {
   ];
 }
 
-function getPosts($slug) {
+function getPosts($slug)
+{
   $query = new WP_Query([
     'post_type' => $slug,
     'posts_per_page' => -1,
@@ -120,7 +125,8 @@ function getPosts($slug) {
   return $query->posts;
 }
 
-function logo($str) {
+function logo($str)
+{
   $logo_data = get_field($str, 'option');
   if (!$logo_data || !isset($logo_data['url'])) {
     return '';
@@ -146,12 +152,9 @@ function logo($str) {
   </a>";
 }
 
-function console($data) {
-  echo '<script>console.log(' . json_encode($data) . ');</script>';
-}
-
 // Function to get services for a specific category
-function get_services_by_category($category_id) {
+function get_services_by_category($category_id)
+{
   return get_posts([
     'post_type' => 'service',
     'posts_per_page' => -1,
@@ -174,13 +177,15 @@ function get_services_by_category($category_id) {
   ]);
 }
 
-function getAssetUrlAcf($str) {
+function getAssetUrlAcf($str)
+{
   $image = get_field($str, 'option');
   $url = isset($image['url']) ? $image['url'] : null;
   return $url;
 }
 
-function get_pages_by_template($template_name) {
+function get_pages_by_template($template_name)
+{
   $pages = get_posts([
     'post_type' => 'page',
     'numberposts' => -1,
